+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### ✨ Improved
 
 * Added `expose --async-readout` flag that finishes the expose command as soon as readout begins.
+* Added `wait-until-idle` command that returns once the spectrographs are idle.
 
 
 ## 0.9.0 - April 13, 2023